Repo // Refactoring the entire LMMgr.
