+++ Source/Modules/SessionCtl_Delegates.swift
@@ -33,22 +33,21 @@ extension SessionCtl: InputHandlerDelegate {
 
   public func performUserPhraseOperation(addToFilter: Bool) -> Bool {
     guard let inputHandler = inputHandler, state.type == .ofMarking else { return false }
-    if !LMMgr.writeUserPhrase(
-      state.data.userPhraseDumped, inputMode: inputMode,
-      areWeDeleting: addToFilter
+    var succeeded = true
+
+    let kvPair = state.data.userPhraseKVPair
+    var userPhrase = LMMgr.UserPhrase(
+      keyArray: kvPair.keyArray, value: kvPair.value, inputMode: inputMode
     )
-      || !LMMgr.writeUserPhrase(
-        state.data.userPhraseDumpedConverted, inputMode: inputMode.reversed,
-        areWeDeleting: addToFilter
-      )
-    {
-      return false
+    if Self.areWeNerfing { userPhrase.weight = -114.514 }
+    LMMgr.writeUserPhrasesAtOnce(userPhrase, areWeFiltering: addToFilter) {
+      succeeded = false
     }
+    if !succeeded { return false }
 
     // 後續操作。
-    let rawPair = state.data.userPhraseKVPair
-    let valueCurrent = rawPair.1
-    let valueReversed = ChineseConverter.crossConvert(rawPair.1)
+    let valueCurrent = userPhrase.value
+    let valueReversed = ChineseConverter.crossConvert(valueCurrent)
 
     // 更新組字器內的單元圖資料。
     // 註：如果已經排除的內容是該讀音下唯一的記錄的話，
@@ -57,10 +56,10 @@ extension SessionCtl: InputHandlerDelegate {
 
     // 因為上述操作不會立即生效（除非遞交組字區），所以暫時塞入臨時資料記錄。
     // 該臨時資料記錄會在接下來的語言模組資料重載過程中被自動清除。
-    let temporaryScore: Double = SessionCtl.areWeNerfing ? -114.514 : 0
     LMMgr.currentLM.insertTemporaryData(
-      keyArray: [rawPair.0], unigram: .init(value: rawPair.1, score: temporaryScore),
-      isFiltering: SessionCtl.areWeNerfing
+      keyArray: userPhrase.keyArray,
+      unigram: .init(value: userPhrase.value, score: userPhrase.weight ?? 0),
+      isFiltering: addToFilter
     )
     // 開始針對使用者半衰模組的清詞處理
     LMMgr.bleachSpecifiedSuggestions(targets: [valueCurrent], mode: IMEApp.currentInputMode)
@@ -179,32 +178,24 @@ extension SessionCtl: CtlCandidateDelegate {
     var succeeded = true
 
     let rawPair = state.candidates[index]
-    let theKey = rawPair.0.joined(separator: InputHandler.keySeparator)
-    let valueCurrent = rawPair.1
-    let valueReversed = ChineseConverter.crossConvert(rawPair.1)
-    let nerfedScore = (action == .toNerf) ? " -114.514" : ""
-    let convertedMark = "#𝙃𝙪𝙢𝙖𝙣𝘾𝙝𝙚𝙘𝙠𝙍𝙚𝙦𝙪𝙞𝙧𝙚𝙙"
-
-    let userPhraseDumped = "\(valueCurrent) \(theKey)\(nerfedScore)"
-    let userPhraseDumpedConverted = "\(valueReversed) \(theKey)\(nerfedScore) \(convertedMark)"
-
-    if !LMMgr.writeUserPhrase(
-      userPhraseDumped, inputMode: inputMode,
-      areWeDeleting: action == .toFilter
+    var userPhrase = LMMgr.UserPhrase(
+      keyArray: rawPair.keyArray, value: rawPair.value, inputMode: inputMode
     )
-      || !LMMgr.writeUserPhrase(
-        userPhraseDumpedConverted, inputMode: inputMode.reversed,
-        areWeDeleting: action == .toFilter
-      )
-    {
+    if action == .toNerf { userPhrase.weight = -114.514 }
+    LMMgr.writeUserPhrasesAtOnce(userPhrase, areWeFiltering: action == .toFilter) {
       succeeded = false
     }
 
+    // 後續操作。
+    let valueCurrent = userPhrase.value
+    let valueReversed = ChineseConverter.crossConvert(valueCurrent)
+
     // 因為上述操作不會立即生效（除非遞交組字區），所以暫時塞入臨時資料記錄。
     // 該臨時資料記錄會在接下來的語言模組資料重載過程中被自動清除。
-    let temporaryScore: Double = (action == .toNerf) ? -114.514 : 0
     LMMgr.currentLM.insertTemporaryData(
-      keyArray: rawPair.0, unigram: .init(value: rawPair.1, score: temporaryScore), isFiltering: action == .toFilter
+      keyArray: userPhrase.keyArray,
+      unigram: .init(value: userPhrase.value, score: userPhrase.weight ?? 0),
+      isFiltering: action == .toFilter
     )
 
     // 開始針對使用者半衰模組的清詞處理
